backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
websockets==12.0
prometheus-client==0.19.0
python-telegram-bot==20.7
aiohttp==3.9.1
httpx==0.25.2
tenacity==8.2.3
celery==5.3.4
python-dotenv==1.0.0
structlog==23.2.0
